+++ src/app/[locale]/(auth)/dashboard/layout.tsx
@@ -11,7 +11,7 @@ export default function DashboardLayout({ children }: DashboardLayoutProps) {
         <main className='relative flex h-full bg-dash-body  w-full flex-row'>
             <Aside />
             <SubMenu />
-            <section className=' flex w-full flex-col gap-5 p-4'>
+            <section className=' flex w-full flex-col gap-5'>
                 {children}
                 <h1 className='text-4xl text-neutral-200'>Dashboard</h1>
                 <div className='h-80 w-full rounded border border-neutral-500/50 bg-neutral-800/20' />
