+++ src/app/[locale]/(unauth)/chats/components/ChatList.tsx
@@ -1,10 +1,11 @@
 import { WebcamIcon } from '@/components/theme/icons'
 import { PhoneIcon } from '@heroicons/react/24/outline'
 import Link from 'next/link'
-import { getChatFiles } from './chat-utils'
+import DashboardLayout from '@/app/[locale]/(auth)/dashboard/_layout'
+import { GetChatFiles } from './chat-utils'
 
 export default async function ChatList() {
-    const chatFiles = await getChatFiles()
+    const chatFiles = await GetChatFiles()
 
     return (
         <aside className='hidden w-64 flex-col border-r bg-card p-6 md:flex'>
@@ -13,15 +14,15 @@ export default async function ChatList() {
                 <h2 className='text-xl font-semibold'>WhatsApp Exports</h2>
             </div>
             <nav className='flex flex-col gap-2 '>
-                {chatFiles.map((file) => (
+                {chatFiles.map((name) => (
                     <Link
-                        key={file}
-                        href={`/chats/${file.replace('-chat.json', '')}`}
+                        key={name}
+                        href={`/chats/${name.replace('-chat.json', '')}`}
                         className='flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-muted'
                         prefetch={false}
                     >
                         <WebcamIcon className='h-5 w-5 text-muted-foreground' />
-                        {file.replace('-chat.json', '')}
+                        {name.replace('-chat.json', '')}
                     </Link>
                 ))}
             </nav>
