+++ src/components/dashboard/guestbook/GuestbookList.tsx
@@ -6,6 +6,7 @@ import { guestbookSchema } from '@/core/models/Schema'
 
 import { EditableGuestbookEntry } from '../../EditableGuestbookEntry'
 import { DeleteGuestbookEntry } from './DeleteGuestbookEntry'
+import { SkeletonGuestbookEntrys } from '@/components/effects/SkeletonLoaders'
 
 interface GuestbookEntry {
     id: number
@@ -16,47 +17,58 @@ interface GuestbookEntry {
 
 export default function GuestbookList() {
     const [entries, setEntries] = useState<GuestbookEntry[]>([]) // Initialize state
-
+    const [isLoading, setIsLoading] = useState(true)
     useEffect(() => {
+        setIsLoading(true)
         const fetchGuestbook = async () => {
             const guestbookData: GuestbookEntry[] = await db
                 .select()
                 .from(guestbookSchema)
                 .all()
-            setEntries(guestbookData) // Update state with fetched data
+            setEntries(guestbookData)
+            setIsLoading(false)
         }
 
-        fetchGuestbook() // Fetch data on component mount
-    }, []) // Empty dependency array means this effect runs once on mount
+        fetchGuestbook()
+    }, [])
 
     return (
         <div
-            className='mt-5 bg-gray-800 py-4 rounded-lg shadow-md'
+            className='mt-5 sm:min-w-[350px] max-w-[350px] py-4 rounded-lg shadow-md'
             data-testid='guestbook-list'
         >
-            {entries.map((entry) => (
+            {isLoading ? (
+
                 <div
-                    key={entry.id}
-                    className='bg-gray-900 p-2 mb-4 rounded-lg gap-2 items-center flex justify-between'
+                    className=' p-2 mb-4  flex-col rounded-lg gap-2 items-center flex justify-between'
                 >
-                    <div className='space-x-4'>
-                        <span className='text-white'>{entry.username}</span>
-                        <span className='text-gray-300'>{entry.body}</span>
-                    </div>
-                    <div className='flex items-center gap-2'>
-                        <div className='text-blue-500 hover:text-blue-700'>
-                            <DeleteGuestbookEntry id={entry.id} />
-                        </div>
-                        <div className='ml-2 text-green-500 hover:text-green-700'>
-                            <EditableGuestbookEntry
-                                id={entry.id}
-                                username={entry.username}
-                                body={entry.body}
-                            />
-                        </div>
-                    </div>
+                    <SkeletonGuestbookEntrys count={7} />
                 </div>
-            ))}
+            ) : (
+                entries.map((entry) => (
+                    <div
+                        key={entry.id}
+                        className='bg-gray-900 p-2 mb-4 rounded-lg gap-2 items-center flex justify-between'
+                    >
+                        <div className='space-x-4'>
+                            <span className='text-white'>{entry.username}</span>
+                            <span className='text-gray-300'>{entry.body}</span>
+                        </div>
+                        <div className='flex items-center gap-2'>
+                            <div className='text-blue-500 hover:text-blue-700'>
+                                <DeleteGuestbookEntry id={entry.id} />
+                            </div>
+                            <div className='ml-2 text-green-500 hover:text-green-700'>
+                                <EditableGuestbookEntry
+                                    id={entry.id}
+                                    username={entry.username}
+                                    body={entry.body}
+                                />
+                            </div>
+                        </div>
+                    </div>
+                ))
+            )}
         </div>
     )
 }
