+++ src/core/data/menu-items.ts
@@ -19,9 +19,10 @@ export const headerDropdownItems: {
         description: 'Just a page to test crud operations.'
     },
     {
-        title: 'Third Tab',
+        title: 'WhatsApp Exports',
         href: '/',
-        description: 'Write some wavy here to get them to click.'
+        description:
+            'Exported chats from WhatsApp fetched from own API, because why not.'
     }
 ]
 
@@ -32,7 +33,7 @@ export const DashboardAsideItems = [
             `<svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="#f7f7f7"><path stroke="#f7f7f7" d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z"/></svg>`
         ),
         hasNotification: false,
-        anchor: 'chats',
+        anchor: 'dashboard/chats',
         isFavourite: false
     },
     {
